Fixed quadtree insertion to handle failed insertions.
Previously any possible sub insertion was allowed, so if insertion was
attempted with a new geometry that overlapped with an existing geometry,
the result would be parts of the object would get inserted, parts would
not. But what we want is for all or none of the geometry to get
inserted. This necessitated checking if insertion is entirely possible
before starting the insertion process (unless I want to implement
rolling back partial insertions).